perbaikan halaman administration bagian teacher

+++ resources/views/livewire/administration.blade.php
@@ -85,10 +85,12 @@
                         class="bg-gray-500 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded">
                             Detail
                         </button>
-                        <button wire:click="deleteConfirmation( {{ $administration->id }} )"
-                                class="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded">
-                            Delete
-                        </button>
+                        @if ($administration->status == 0)
+                            <button wire:click="deleteConfirmation( {{ $administration->id }} )"
+                                    class="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded">
+                                Delete
+                            </button>
+                        @endif
                     </td>
                 </tr>
             @empty
